requirements.txt
apturl==0.5.2
asgiref==3.2.3
asn1crypto==0.24.0
bcrypt==3.1.6
beautifulsoup4==4.8.0
blinker==1.4 
certifi==2018.8.24
chardet==3.0.4 
cookies==2.2.1
cryptography==2.6.1
cupshelpers==1.0
dbus-python==1.2.12
defer==1.0.6
distro==1.3.0 
Django==2.1.3
django-crispy-forms==1.7.2
django-model-utils==3.2.0 
entrypoints==0.3
fasteners==0.12.0
future==0.18.2
html5lib==1.0.1
httplib2==0.11.3
idna==2.6
keyring==18.0.1
keyrings.alt==3.1.1
language-selector==0.1
launchpadlib==1.10.7
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libvirt-python==5.0.0
lockfile==0.12.2 
lxml==4.4.1 
Mako==1.0.7
MarkupSafe==1.1.0
monotonic==1.5
mysqlclient==1.4.6
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
openshot-qt==2.4.3
paramiko==2.6.0
pexpect==4.6.0
Pillow==5.3.0
protobuf==3.6.1
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73 
PyJWT==1.7.0
pymacaroons==0.13.0
PyNaCl==1.3.0
PyQt5==5.12.3
pyRFC3339==1.1 
python-dateutil==2.7.3
python-debian==0.1.36
python-docx==0.8.10
python-twitter==3.5
pytz==2018.4
pyxdg==0.25
PyYAML==5.1.2
pyzmq==17.1.2
reportlab==3.5.23
requests==2.21.0
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
scour==0.37
SecretStorage==2.3.1
simplejson==3.16.0
sip==4.19.18
six==1.12.0
soupsieve==1.9.2
sqlparse==0.2.4
system-service==0.3
systemd-python==234
ubuntu-advantage-tools==19.5
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.24.1
usb-creator==0.3.7
wadllib==1.3.3
webencodings==0.5.1
xkit==0.0.0
djangorestframework
